refactor magnitude stuff into separate service
