+++ Commands/Implementations/Terminate.cs
@@ -1,5 +1,4 @@
-﻿using obs_cli.Data;
-using obs_cli.Enums;
+﻿using obs_cli.Enums;
 using obs_cli.Services;
 using System;
 using System.Collections.Generic;
@@ -17,10 +16,8 @@ namespace obs_cli.Commands.Implementations
 
         public override void Execute()
         {
-            Store.Data.Audio.InputMeter.RemoveCallback();
-            Store.Data.Audio.OutputMeter.RemoveCallback();
+            MagnitudeService.Teardown();
             PipeService.Teardown();
-
             Environment.Exit(0);
         }
     }
