+++ Services/AudioService.cs
@@ -113,7 +113,7 @@ namespace obs_cli.Services
             Store.Data.Audio.InputMeter = new VolMeter();
             Store.Data.Audio.InputMeter.Level = float.NegativeInfinity;
             Store.Data.Audio.InputMeter.AttachSource(Store.Data.Audio.InputSource);
-            Store.Data.Audio.InputMeter.AddCallBack(InputVolumeCallback);
+            Store.Data.Audio.InputMeter.AddCallBack(MagnitudeService.InputVolumeCallback);
 
             List<AudioDevice> allAudioInputs = GetAudioInputDevices();
             bool savedIsInAvailableInputs = allAudioInputs.Any(x => x.id == savedAudioInputId);
@@ -161,7 +161,7 @@ namespace obs_cli.Services
             Store.Data.Audio.OutputMeter = new VolMeter();
             Store.Data.Audio.OutputMeter.Level = float.NegativeInfinity;
             Store.Data.Audio.OutputMeter.AttachSource(Store.Data.Audio.OutputSource);
-            Store.Data.Audio.OutputMeter.AddCallBack(OutputVolumeCallback);
+            Store.Data.Audio.OutputMeter.AddCallBack(MagnitudeService.OutputVolumeCallback);
 
             List<AudioDevice> allAudioOutputs = GetAudioOutputDevices();
             bool savedIsInAvailableOutputs = allAudioOutputs.Any(x => x.id == savedAudioOutputId);
@@ -180,42 +180,6 @@ namespace obs_cli.Services
             }
 
             return usedAudioOutputId;
-        }
-
-        private static Timer _magnitudeTimer;
-
-        public static void SetupMagnitudeCycle()
-        {
-            _magnitudeTimer = new Timer();
-            _magnitudeTimer.Interval = 33;
-            _magnitudeTimer.Elapsed += SendMagnitudes;
-            _magnitudeTimer.Start();
-        }
-
-        private static void SendMagnitudes(object sender, ElapsedEventArgs e)
-        {
-            var magnitudes = new AudioMagnitudesResponse
-            {
-                IsSuccessful = true,
-                AudioInputLevel = Store.Data.Audio.InputMeter.Level,
-                AudioOutputLevel = Store.Data.Audio.OutputMeter.Level
-            };
-
-            Store.Data.Pipe.Magnitude.PushMessage(new Message { Text = new JavaScriptSerializer().Serialize(magnitudes) });
-        }
-
-        // As of OBS 21.0.1, audo meters have been reworked. We now need to calculate and draw ballistics ourselves. 
-        // Relevant commit: https://github.com/obsproject/obs-studio/commit/50ce2284557b888f230a1730fa580e82a6a133dc#diff-505cedf4005a973efa8df1e299be4199
-        // This is probably an over-simplified calculation.
-        // For practical purposes, we are treating -60 as 0 and -9 as 1.
-        private static void InputVolumeCallback(IntPtr data, float[] magnitude, float[] peak, float[] input_peak)
-        {
-            Store.Data.Audio.InputMeter.Level = magnitude[0];
-        }
-
-        private static void OutputVolumeCallback(IntPtr data, float[] magnitude, float[] peak, float[] input_peak)
-        {
-            Store.Data.Audio.OutputMeter.Level = magnitude[0];
         }
 
         /// <summary>
